chainbackends: extract bitcoind package submitter into own package
The direct-to-bitcoind JSON-RPC PackageSubmitter was originally
added under the harness package for itest wiring. That same helper
is now also wired into the production daemon from cmd/darepod when
the operator supplies bitcoind.{host,user,pass} flags, so a test-
only package is the wrong home: it pulls harness imports into
production binaries and muddles the boundary between test
infrastructure and shipped code.

Move the type into its own chainbackends/bitcoindrpc sub-package
with a narrower name (PackageSubmitter instead of
BitcoindPackageSubmitter) and constructor (New). The behavior is
unchanged: same 30s backstop timeout, same JSON-RPC envelope
handling, same maxfeerate=0 override for CPFP child acceptance.
cmd/darepod switches to the new import and the old copy is
removed.
