+++ harness/AGENTS.md
@@ -8,10 +8,6 @@ containers with network isolation for end-to-end testing.
 ## Key Types
 
 - `Harness` — Top-level test harness owning bitcoind, lnd, and arkd lifecycle.
-- `BitcoindPackageSubmitter` — Implements `chainbackends.PackageSubmitter` via
-  direct bitcoind JSON-RPC. Injected into `darepod.Config.PackageSubmitter` to
-  enable v3 CPFP package submission in integration tests without going through
-  LND. Constructor: `NewBitcoindPackageSubmitter(host, user, password)`.
 - `LndInstance` — Manages an LND container's lifecycle and connection.
 - `TapdHarness` — Optional Tapd instance for asset-related tests.
